Fix build under Linux/GCC Windows/MSVC (#3502)
* Avoid unconditionally setting C++ flags

These were already handled via CMAKE_C_FLAGS and CMAKE_CXX_FLAGS and
this breaks when compiling plain C under GCC, which considers unknown
options an error when compiling with -Werror.

* Fix compilation in Linux/GCC and also Windows/MSVC

GCC and MSVC aren't happy about `using Query = testutil::Query;` when
there's already `class Query` in scope.

* Fix objc_class.h in Windows

* Fix warnings that have crept in on Windows

+++ Firestore/core/src/firebase/firestore/model/transform_operations.cc
@@ -155,7 +155,7 @@ double AsDouble(const FieldValue& value) {
   if (value.type() == FieldValue::Type::Double) {
     return value.double_value();
   } else if (value.type() == FieldValue::Type::Integer) {
-    return value.integer_value();
+    return static_cast<double>(value.integer_value());
   } else {
     HARD_FAIL("Expected 'operand' to be of numeric type, but was %s (type %s)",
               value.ToString(), value.type());
